test(CH): split tests into modules
