+++ frame/clearing-house/src/tests/add_margin.rs
@@ -1,0 +1,60 @@
+use crate::{
+	mock::{
+		accounts::ALICE,
+		assets::{PICA, USDC},
+		runtime::{Balance, ExtBuilder, Origin, Runtime, System as SystemPallet, TestPallet},
+	},
+	pallet::{AccountsMargin, Error, Event},
+	tests::run_to_block,
+};
+
+use frame_support::{assert_noop, assert_ok};
+use orml_tokens::Error as TokenError;
+
+// ----------------------------------------------------------------------------------------------------
+//                                             Add Margin
+// ----------------------------------------------------------------------------------------------------
+
+#[test]
+fn add_margin_returns_transfer_error() {
+	ExtBuilder::default().build().execute_with(|| {
+		let origin = Origin::signed(ALICE);
+		assert_noop!(
+			TestPallet::add_margin(origin, USDC, 1_000_u32.into()),
+			TokenError::<Runtime>::BalanceTooLow
+		);
+	});
+}
+
+#[test]
+fn deposit_unsupported_collateral_returns_error() {
+	ExtBuilder { balances: vec![(ALICE, PICA, 1_000_000)], ..Default::default() }
+		.build()
+		.execute_with(|| {
+			let origin = Origin::signed(ALICE);
+			assert_noop!(
+				TestPallet::add_margin(origin, PICA, 1_000_u32.into()),
+				Error::<Runtime>::UnsupportedCollateralType
+			);
+		});
+}
+
+#[test]
+fn deposit_supported_collateral_succeeds() {
+	ExtBuilder { balances: vec![(ALICE, USDC, 1_000_000)], ..Default::default() }
+		.build()
+		.execute_with(|| {
+			run_to_block(1);
+			let account = ALICE;
+			let asset = USDC;
+			let amount: Balance = 1_000_u32.into();
+
+			let before = AccountsMargin::<Runtime>::get(&account).unwrap_or_default();
+			assert_ok!(TestPallet::add_margin(Origin::signed(account), asset, amount));
+
+			SystemPallet::assert_last_event(Event::MarginAdded { account, asset, amount }.into());
+
+			let after = AccountsMargin::<Runtime>::get(&account).unwrap_or_default();
+			assert_eq!(after - before, amount);
+		})
+}
